% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm.R
\name{terms.BTM}
\alias{terms.BTM}
\title{Get highest token probabilities for each topic}
\usage{
\method{terms}{BTM}(x, threshold = +Inf, top_n = 5, ...)
}
\arguments{
\item{x}{an object of class BTM as returned by \code{\link{BTM}}}

\item{threshold}{threshold in 0-1 range. Only the terms which are more likely than the threshold are returned for each topic}

\item{top_n}{integer indicating to return the top_n}

\item{...}{not used}
}
\value{
a list of data.frames where each data.frame contains token and probability ordered from high to low.
The list is the same length as the number of topics.
}
\description{
Get highest token probabilities for each topic
}
\examples{
library(udpipe)
data("brussels_reviews_anno", package = "udpipe")
x <- subset(brussels_reviews_anno, language == "nl")
x <- subset(x, xpos \%in\% c("NN", "NNP", "NNS"))
model  <- BTM(x, k = 5, iter = 5, trace = TRUE)
terms(model)
}
